ECU1370: set the serial console to ttymxc3
The serial console is defined in the device tree to be ttymxc3, but the
machine configuration uses ttymxc0. Turns out this creates an issue when
this variable is used to let getty know where to start a userspace
console, which is really ttymxc3.

On Yocto's core-image-minimal Getty starts from both ttymxc0 and
ttymxc3, so this error was not noticed.

Changelog-entry: Fix: set ttymxc3 as the serial console.
Signed-off-by: Lisandro Pérez Meyer <lpmeyer@ics.com>